+++ CsharpExtras/Enumerable/OneBased/OneBasedArray.cs
@@ -61,7 +61,7 @@ namespace CsharpExtras.Enumerable.OneBased
 
         public (int index, TVal element)? FindFirstOccurrenceOfSet(ISet<TVal> set)
         {
-            return FindFirstOccurrenceOfSet(set, 1, Length + 1);
+            return FindFirstOccurrenceOfSet(set, 1, Length);
         }
 
         public int FirstIndexOf(Func<TVal, bool> matchFunction)